% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{endogCopula-package}
\alias{endogCopula}
\alias{endogCopula-package}
\title{endogCopula: Gaussian Copula Based Endogeneity Corrections}
\description{
The endogCopula package implements a collection of copula-based estimators
for addressing endogeneity in linear models. It includes cross-sectional
estimators from Park and Gupta (2012), Yang et al. (2025), Hu et al. (2025),
Breitung et al. (2024), Haschka (2024), and Liengaard et al. (2025).
}
\author{
\strong{Maintainer}: Ashwin Malshe \email{ashwin@malshe.com}

}
\keyword{internal}
