% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.endog_copula_fit}
\alias{print.endog_copula_fit}
\title{Print method for \code{endog_copula_fit}}
\usage{
\method{print}{endog_copula_fit}(x, ...)
}
\arguments{
\item{x}{An object returned by one of the copula estimators.}

\item{...}{Passed to downstream methods.}
}
\description{
Print method for \code{endog_copula_fit}
}
