% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.endog_copula_fit}
\alias{summary.endog_copula_fit}
\title{Summary method for \code{endog_copula_fit}}
\usage{
\method{summary}{endog_copula_fit}(object, ...)
}
\arguments{
\item{object}{An object returned by one of the copula estimators.}

\item{...}{Passed to downstream methods.}
}
\description{
Summary method for \code{endog_copula_fit}
}
